ActiveDecorator を導入し、モデルのビュー関連ロジックをデコレータに移行
- Gemfile に active_decorator を追加し、bundle install を実行
- article_decorator.rb と user_decorator.rb を新規作成
- 各モデルに含まれていたビュー向けメソッドを対応するデコレータに移動
