+++ Gemfile
@@ -20,6 +20,7 @@ gem "jbuilder"
 gem 'faker'
 gem 'hamlit'
 gem 'annotate'
+gem 'active_decorator'
 gem 'better_errors'
 gem 'binding_of_caller'
 gem 'devise'
