+++ app/models/article.rb
@@ -31,9 +31,7 @@ class Article < ApplicationRecord
     I18n.l(created_at, format: :default)
   end
 
-  def author_name
-    user.display_name
-  end
+  
 
   def like_count
     likes.count
